# docs/static/windrunner-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Windrunner API",
    "description": "API-key authenticated endpoints for Windrunner integrations.\n\nSend the key as `Authorization: Bearer <key>`. Each operation\nrequires the scope named in the API reference. Project-scoped\nreads and writes also require the key owner to have access to\nthat project. Team and team-project directory endpoints are\norganization-wide; audit-log endpoints require an admin or\nsuperadmin API-key owner.\n\nCollection endpoints use zero-based `page` and `size` parameters.\nThe maximum page size is 100 and pagination metadata is returned\nin the standard `meta` response object. Search responses are\nbounded by their `limit` parameter (maximum 100).\n",
    "version": "v1"
  },
  "security": [
    {
      "ApiKeyBearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects/{projectId}/content-order": {
      "put": {
        "tags": [
          "Project content"
        ],
        "operationId": "reorder",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "getProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/search": {
      "get": {
        "tags": [
          "Project content"
        ],
        "operationId": "search",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "WorkItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "parentWorkItemId": {
            "type": "string"
          },
          "sortIndex": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "priority": {
            "type": "string"
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "default": null,
        "description": "A single error reported for a request.",
        "properties": {
          "code": {
            "type": "string",
            "default": "",
            "description": "Stable machine-readable error code.",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "default": "",
            "description": "Human-readable error message."
          },
          "field": {
            "type": "string",
            "default": "",
            "description": "Request field the error relates to, when applicable."
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "default": "",
            "description": "Additional structured details for this error."
          }
        }
      },
      "ApiMeta": {
        "type": "object",
        "default": null,
        "description": "Request metadata. Pagination fields are present only on paginated responses.",
        "properties": {
          "requestId": {
            "type": "string",
            "default": "",
            "description": "Server-generated request identifier, useful for support."
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "default": "",
            "description": "Current page number, zero-based."
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "default": "",
            "description": "Page size used for this response."
          },
          "totalItems": {
            "type": "integer",
            "format": "int64",
            "default": "",
            "description": "Total number of items across all pages."
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "default": "",
            "description": "Total number of pages."
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "default": null,
        "description": "Standard response envelope for all Windrunner API endpoints.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WorkItem",
            "default": "",
            "description": "The requested resource. Null when errors is non-empty."
          },
          "errors": {
            "type": "array",
            "default": "",
            "description": "Errors reported for this request. Empty on success.",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta",
            "default": "",
            "description": "Request metadata. Present on paginated responses."
          }
        }
      },
...
      "ContentOrderItemRef": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          }
        }
      },
      "ContentReorderRequest": {
        "type": "object",
        "properties": {
          "parentWorkItemId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentOrderItemRef"
            }
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ownerUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ownerDisplayNames": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyBearerAuth": {
        "type": "http",
        "description": "Use an active Windrunner API key. Create and revoke keys from My Account \u2192 API keys.",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    }
  }
}